refactor: 요청, 응답에 대한 패키지 분리
