+++ tomcat/src/main/java/nextstep/jwp/presentation/IndexController.java
@@ -1,11 +1,11 @@
 package nextstep.jwp.presentation;
 
-import org.apache.coyote.http11.RequestReader;
-import org.apache.coyote.http11.Response;
+import org.apache.coyote.http11.request.RequestReader;
+import org.apache.coyote.http11.response.Response;
 
 import java.io.IOException;
 
-import static org.apache.coyote.http11.StatusCode.OK;
+import static org.apache.coyote.http11.response.StatusCode.OK;
 
 public class IndexController implements Controller {
