+++ tomcat/src/main/java/nextstep/jwp/presentation/LoginController.java
@@ -2,18 +2,18 @@ package nextstep.jwp.presentation;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.RequestReader;
-import org.apache.coyote.http11.Response;
-import org.apache.coyote.http11.SessionManager;
+import org.apache.coyote.http11.request.RequestReader;
+import org.apache.coyote.http11.response.Response;
+import org.apache.coyote.http11.response.SessionManager;
 
 import java.io.IOException;
 
-import static org.apache.coyote.http11.Header.LOCATION;
+import static org.apache.coyote.http11.response.Header.LOCATION;
 import static org.apache.coyote.http11.Method.GET;
 import static org.apache.coyote.http11.Method.POST;
-import static org.apache.coyote.http11.StatusCode.FOUND;
-import static org.apache.coyote.http11.StatusCode.OK;
-import static org.apache.coyote.http11.StatusCode.UNAUTHORIZED;
+import static org.apache.coyote.http11.response.StatusCode.FOUND;
+import static org.apache.coyote.http11.response.StatusCode.OK;
+import static org.apache.coyote.http11.response.StatusCode.UNAUTHORIZED;
 
 public class LoginController implements Controller {
