+++ tomcat/src/main/java/nextstep/jwp/presentation/MainController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.presentation;
 
-import org.apache.coyote.http11.RequestReader;
-import org.apache.coyote.http11.Response;
-import org.apache.coyote.http11.StatusCode;
+import org.apache.coyote.http11.request.RequestReader;
+import org.apache.coyote.http11.response.Response;
+import org.apache.coyote.http11.response.StatusCode;
 
 public class MainController implements Controller{
