+++ tomcat/src/main/java/nextstep/jwp/presentation/OtherController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.presentation;
 
-import org.apache.coyote.http11.RequestReader;
-import org.apache.coyote.http11.Response;
-import org.apache.coyote.http11.StatusCode;
+import org.apache.coyote.http11.request.RequestReader;
+import org.apache.coyote.http11.response.Response;
+import org.apache.coyote.http11.response.StatusCode;
 
 import java.io.IOException;
