+++ tomcat/src/main/java/org/apache/coyote/http11/response/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.response;
 
 import nextstep.jwp.model.User;
